Formatted code, still need to be retested on the real robot
