+++ speech_recognition_server/action/SpeechRecognitionActivated.action
@@ -1,0 +1,10 @@
+#goal definition
+string language
+bool skip_keyword #Start recognizing immediately, without waiting for keyword
+bool tell_back   #Robot tell back what it just heard
+---
+---
+#Each time a keyword is detected and speech is recognized , a new vector is sent
+#Each time a keyword is detected and speech is recognized, a new vector is sent
+#with the recognitions
+string[] recognition_results
